added upsample layers

+++ mirnetv2/model/downsample.py
@@ -16,7 +16,7 @@ class DownBlock(tf.keras.layers.Layer):
 
 class DownSampleBlock(tf.keras.layers.Layer):
     def __init__(
-        self, channels: int, scale_factor: int, channel_factor: int, *args, **kwargs
+        self, channels: int, scale_factor: int, channel_factor: float, *args, **kwargs
     ):
         super(DownSampleBlock, self).__init__(*args, **kwargs)
         layers = []
